Fix some timing issues
The interval for switching directions in figure8 mode should have its own enum. It should not depend on whether the facing-player fix is enabled.

Set the shoot delay to 291 ticks regardless of whether the facing-player fix is enabled. If the delay is halved, he shoots more frequently and becomes very hard to fight.

In both cases, we are sticking with the DOS values