Updates to run on macOS (see description)
This is based on two things.

PR #473 from @apancik
https://github.com/luciusDXL/TheForceEngine/pull/473

The comment from @gilmorem560 that updated AngelScript code resolves compilation issues on Apple Silicon
https://github.com/luciusDXL/TheForceEngine/issues/90#issuecomment-2816098327
https://www.angelcode.com/angelscript/

The code changes are sort of like a manual rebase of #473 so perhaps a proper rebase of that and the manual grafting of AngelScript would be a better solution. In any event I want to give credit to the people who really did the work.